First version of the Pascal's triangle II problem.
Adapted the  mathematical approach, instead of the iterative solution. The time/space complexity is relatively optimal, though, it suffers from system upperbound limitation.

Still there is space to optimize. E.g. the combination function for the last half  is not neccessary, we should be able to make use of the symmentry.